Update html_lists.html
we can also select what type of bulleting or numbering patterns we can use in our lists. For example the ordered list will have 1,2,3,4, or a,b,c,d or A,B,C,D or I,II,III,IV etc like this.Similarly for unordered list the bullets design can be modified.